SQB/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
wsproto>=1.2.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
python-dotenv>=1.0.1
groq>=0.9.0
pydantic>=2.10.0
httpx>=0.27.0
